fix bug for propagating component issue to cluster
